Fix problem with ordering of Koans in AboutLambdas
There are two Koans that jump to the front of the line and don't get
executed in order like the rest.  The problem seems to be in the way
they are named.  My best guess is that Java reflection is ordering them
first because a member variable with the same name comes above it.
However, adding characters to the end of the name doesn't fix the
problem, you have the change something else.  It's possible that this is
a problem with Java reflection and there isn't anything else that can be
done, but I suspect there might be a bug in the Koans library that could
explain the odd behavior.  Fixing that could help avoid similar issues
elsewhere.