Possible fix for Issue 79 - use of assertNotSame.
[Issue 79](https://github.com/matyb/java-koans/issues/79) questions the
need for the `AboutAssertions.assertNotSameInstance` method, specifically
the `aboutNotSame` assertion.

This is my attempt to make both the positive assertion in
`assertSameInstance` and the negative assertion more clear (if much
longer) by making them more similar, and by giving each an example of
how equality is different from "sameness".

I'm still not sure it justifies the existence of `aboutNotSame`, which
is used literally nowhere else.